scripts/bible/requirements.txt
requests>=2.31
anthropic>=0.39
python-dotenv>=1.0
